added comments more comments and fixed up paths being made from corner to corner in wall space
