+++ inc/a_star.h
@@ -1,8 +1,8 @@
 #ifndef A_STAR_H
 # define A_STAR_H
 
-# define DELTAS_LEN 8
-# define DELTAS {{1, 0}, {0, 1}, {-1, 0}, {0, -1}, {-1, -1}, {-1, 1}, {1, -1}, {1, 1}}
+# define DELTAS_LEN 4
+#define DELTAS {{0, -1}, {0, 1}, {-1, 0}, {1, 0}}
 
 typedef struct s_game t_game;
 
@@ -19,6 +19,7 @@ void    star_deallocate_node(t_star_node **node);
 void    star_insert_node(t_game *game, t_star_node **node, t_star_node *parent, t_vec2 pos);
 t_vec2	*star_construct_path(t_star_node *end_node, t_star_node **open, t_star_node **close);
 t_vec2	*star_find_path(t_game *game, t_vec2 start, t_vec2 target);
+void	star_get_neighbors(t_game *game, t_star_node *curr, t_star_node **open, t_star_node **close);
 void	print_vector_path(t_game *game);
 
 #endif
